Fix files with short paths crashing the emulator
The file path being loaded is dynamically replaced
with an emulator-local copy stored in a SAVE
directory, which allows the original disk data to
remain unmodified.

In case the original file path is shorter than the
destination "save" path, the file name buffer was
undersized in some situations, which could cause a
crash on load.